change package and default to nil logger
